Storage Implementation
This commit implements the Storage struct. The Storage struct
is responsible for data persistence in this project.

Minor improvements to main.
Minor improvements to project structure.

Changed the default error handling from io::Result (bad thing to do)
to Box<dyn Error>, not the best solution but good enough for now.
The error.rs handles that crap.